move StatisticsServiceImpl to rest service package
